+++ malpolon/models/custom_models/glc2024_pre_extracted_prediction_system.py
@@ -15,9 +15,8 @@ from torch import Tensor, nn
 from torch.optim.lr_scheduler import CosineAnnealingLR
 from torchvision import models
 
-from malpolon.models import ClassificationSystem
-
-from .utils import check_optimizer
+from malpolon.models.standard_prediction_systems import ClassificationSystem
+from malpolon.models.utils import check_optimizer
 
 
 class ClassificationSystemGLC24(ClassificationSystem):
@@ -35,7 +34,9 @@ class ClassificationSystemGLC24(ClassificationSystem):
         metrics: Optional[dict[str, Callable]] = None,
         task: str = 'classification_binary',
         loss_kwargs: Optional[dict] = {},
-        hparams_preprocess: bool = True
+        hparams_preprocess: bool = True,
+        download_weights: bool = None,
+        weights_dir: str = 'outputs/glc24_cnn_multimodal_ensemble/',
     ):
         if isinstance(loss_kwargs, omegaconf.dictconfig.DictConfig):
             loss_kwargs = OmegaConf.to_container(loss_kwargs, resolve=True)
@@ -45,6 +46,12 @@ class ClassificationSystemGLC24(ClassificationSystem):
         super().__init__(model, lr, weight_decay, momentum, nesterov, metrics, task, loss_kwargs, hparams_preprocess)
         optimizer = torch.optim.AdamW(model.parameters(), lr=lr)
         self.optimizer = check_optimizer(optimizer)
+        if download_weights:
+            self.download_weights("https://lab.plantnet.org/seafile/f/5e9fda3acdbb4561bb30/?dl=1",
+                                  weights_dir,
+                                  filename="2024-06-24_19-14-48.zip",
+                                  md5="53e17e7e09834f5bf29c8e5e11e2e304")
+
 
     def configure_optimizers(self):
         """Override default optimizer and scheduler.
@@ -95,54 +102,3 @@ class ClassificationSystemGLC24(ClassificationSystem):
     def predict_step(self, batch, batch_idx, dataloader_idx=0):  # noqa: D102 pylint: disable=C0116
         x_landsat, x_bioclim, x_sentinel, y, survey_id = batch
         return self(x_landsat, x_bioclim, x_sentinel)
-
-
-class MultimodalEnsemble(nn.Module):
-    """Multimodal ensemble model processing Sentinel-2A, Landsat & Bioclimatic data.
-
-    Inherits torch nn.Module.
-    """
-    def __init__(self, num_classes=11255, positive_weigh_factor=1.0, **kwargs):
-        super().__init__(**kwargs)
-        self.positive_weigh_factor = positive_weigh_factor
-
-        self.landsat_model = models.resnet18(weights=None)
-        self.landsat_norm = nn.LayerNorm([6, 4, 21])
-        # Modify the first convolutional layer to accept 6 channels instead of 3
-        self.landsat_model.conv1 = nn.Conv2d(6, 64, kernel_size=3, stride=1, padding=1, bias=False)
-        self.landsat_model.maxpool = nn.Identity()
-
-        self.bioclim_model = models.resnet18(weights=None)
-        self.bioclim_norm = nn.LayerNorm([4, 19, 12])
-        # Modify the first convolutional layer to accept 4 channels instead of 3
-        self.bioclim_model.conv1 = nn.Conv2d(4, 64, kernel_size=3, stride=1, padding=1, bias=False)
-        self.bioclim_model.maxpool = nn.Identity()
-
-        self.sentinel_model = models.swin_t(weights="IMAGENET1K_V1")
-        # Modify the first layer to accept 4 channels instead of 3
-        self.sentinel_model.features[0][0] = nn.Conv2d(4, 96, kernel_size=(4, 4), stride=(4, 4))
-        self.sentinel_model.head = nn.Identity()
-
-        self.ln1 = nn.LayerNorm(1000)
-        self.ln2 = nn.LayerNorm(1000)
-        self.fc1 = nn.Linear(2768, 4096)
-        self.fc2 = nn.Linear(4096, num_classes)
-
-        self.dropout = nn.Dropout(p=0.1)
-
-    def forward(self, x, y, z):  # noqa: D102 pylint: disable=C0116
-        x = self.landsat_norm(x)
-        x = self.landsat_model(x)
-        x = self.ln1(x)
-
-        y = self.bioclim_norm(y)
-        y = self.bioclim_model(y)
-        y = self.ln2(y)
-
-        z = self.sentinel_model(z)
-
-        xyz = torch.cat((x, y, z), dim=1)
-        xyz = self.fc1(xyz)
-        xyz = self.dropout(xyz)
-        out = self.fc2(xyz)
-        return out
