+++ malpolon/models/model_builder.py
@@ -17,11 +17,16 @@ import timm
 from torch import nn
 from torchvision import models
 
+from malpolon.models.custom_models.glc2024_multimodal_ensemble_model import \
+    MultimodalEnsemble
+
 if TYPE_CHECKING:
     from typing import Any, Callable, Optional
 
     Provider = Callable[..., nn.Module]
     Modifier = Callable[..., nn.Module]
+
+MALPOLON_MODELS = {'glc24_cnn_multimodal_ensemble': MultimodalEnsemble,}
 
 
 class _ModelBuilder:
@@ -152,6 +157,27 @@ def timm_model_provider(
         )
     return model
 
+# def malpolon_model_provider(
+#     model_name: str, *model_args: Any, **model_kwargs: Any
+# ) -> nn.Module:
+#     """Return a model from Malpolon's models list.
+
+#     This method uses Malpolon's internal model listing to retrieve a
+#     model.
+
+#     Parameters
+#     ----------
+#     model_name : str
+#         name of the model to retrieve from torchvision's library
+
+#     Returns
+#     -------
+#     nn.Module
+#         model object
+#     """
+#     model = MALPOLON_MODELS[model_name]
+#     model = model(*model_args, **model_kwargs)
+#     return model
 
 def _find_module_of_type(
     module: nn.Module, module_type: type, order: str
@@ -306,6 +332,7 @@ ModelBuilder = _ModelBuilder()
 
 ModelBuilder.register_provider("torchvision", torchvision_model_provider)
 ModelBuilder.register_provider("timm", timm_model_provider)
+# ModelBuilder.register_provider("malpolon", malpolon_model_provider)
 
 ModelBuilder.register_modifier(
     "change_first_convolutional_layer",
